+++ taichi/llvm/struct_llvm.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#ifdef TI_WITH_LLVM
 // Codegen for the hierarchical data structure (LLVM)
 #include "taichi/llvm/llvm_program.h"
 #include "taichi/llvm/llvm_codegen_utils.h"
@@ -56,5 +55,3 @@ class StructCompilerLLVM : public StructCompiler, public LLVMModuleBuilder {
 
 }  // namespace lang
 }  // namespace taichi
-
-#endif  //#ifdef TI_WITH_LLVM
